Restructures macro to check for existing forecasts before calling
The new version of the macro now checks for existing notes before
it ever calls Haarpist, which streamlines the app and prevents
unnecessary calls to Haarpist. Now, we can guarantee that every
call to Haarpist will result in meaningful data that will be used.

It isn't yet finished - it still needs to logic for fetching and
posting the forecast.