Adds new forecasting macros and improves existing macro
This adds macros to post existing forecast notes from the
`SimpleCalendar` macro (if any) and to blindly fetch a Haarpist
forecast without interacting with any calendar modules, so as to
not require any as a dependency.

It also changes the way the macros map setting cities to
real-world cities for the purpose of fetching weather data from
Haarpist.

It now keeps a map of setting cities as the key, with the values
being the matching supported real-world city at the top of the
macro for easy editing on the fly.