Adds blast & kill template macro
This macro places a measure template for each of the provided
blast and kill values, centering them on a token named "Boom"
which must be present on the scene. If "Boom" is not on the scene,
an error is displayed to the user.

In the future, I would like for the system to draw a template
preview and place it wherever the user clicks rather than having
to put a specific Actor on the Scene.